central/requirements.txt
kafka-python
mysql-connector-python
python-dotenv